Core migration: Cryptography.PasswordHash
